Show '...' instead of negative countdown
If `timeLeft` is less than 0, the countdown likely continued and is
waiting for the interval to clean up this tab. It's also possible the
number of tabs is not below `minTabs`, which has stopped the countdown
and locked this at a negative `timeLeft` until another tab is opened to
jump start the countdown again.

Fixes #110