Remove underutilized update notifications
This feature historically notified users of new fixes and features, but
it hasn't been used in a long time. It is abnormal for Chrome Extensions
to request notifications permissions without doing something with
notifications (anecdotally). Nuke the permission, nuke its usage.

This removes the `"notifications"` permission from the extension
manifest.