Hoist image loading to move loading after mount
Rather than check for loading images after a single, global 1s timeout,
start the 1s timeout after each mount of the `CorralTab`. This ensures
that the N calls to `getBoundingClientRect` and the subsequent many
re-layouts are done after the popup is fully rendered. This makes
switching back and forth to the Corral Tab responsive to the user's
actions without any delay for re-layouts.

Fixes #163