Load favicons from cache for super speedy popup
The popup now makes no network requests when it opens. The popup is
noticeably faster when it opens.

This uses a Chrome-specific API that might not be available in other
browsers like Firefox. In order to ship Firefox (#185) and Safari (#194)
versions, image lazy loading might need to be re-added. Even if that is
the case, this code should remain because the impact on the experience
is significant.

Fixes #197