Preserve tab timers across browser restart
Persist tab close timers to local storage so they can be read from there
during initialization. If a tab's timer doesn't exist in local storage
during init, it gets a new timer with the max allowable time.

This required a lengthy rework of the background initialization to
ensure async reading happened in the correct order and the tab times
could be deterministically read.

Fixes #130