Unmount Popup component to clean up subscriptions
This was painfully difficult to debug and equally painfully obvious once
I figured it out. The popup needs to clean up its subscriptions since
the background page lives forever. When the popup is going away, the
window fires the [`pagehide` event][0]. It does *not* fire
`beforeunload` or `unload`, although I discovered that through testing
and not through any documentation.

[0]: https://developer.mozilla.org/en-US/docs/Web/Events/pagehide