Import in background.js to prevent Firefox cleanup
When objects are allocated in the context of an extension's popup, they
are de-allocated when the popup disappears. Subsequently accessing those
objects results in a [`TypeError: can't access dead object`][0] error.
To ensure Firefox doesn't clean up the imported tab data, do the
importing from the background page instead of from the popup.

* Convert import/export stuff to Redux actions so no goofy binding to
  the store needs to be done (that's effectively what Redux gets you)

[0]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Dead_object