Pass tabs around by reference; IDs aren't unique
Tab IDs are unique only within a given session. After that session is
gone, like when a tab is closed, the IDs can be reused. Tabs must be
passed around by reference and not by ID because the IDs are not
guaranteed to be unique and in practice most certainly are *not* unique.

Addresses #216