Update filterTabGroups visibility logic.
Previously, we'd explicitly look for the browser. This change updates the logic to check if groupId is available on tabs
and shows the setting if so.

This will mean that filtering on tab groups will just start working if Firefox adds a similar API.

This change doesn't remove the setting of the global BROWSER webpack variable.