Turn singletons into classes, enable dep injection
These singletons all call into each other, which makes it impossible to
split the background page (soon to be service worker) from the popup.
The popup calling into a singleton is actually a message that should go
to the service worker.

Turning these singletons into classes enables instantiating them both
in the service worker env and in the browser/popup env.