Fix unwrangling tabs; alias action to background script
Because the "unwrangle tabs" action calls `chrome` APIs but also likely
forces the popup to close, it is likely the popup is closed before
state cleanup is completed. Instead, alias the action and perform the
work in the background script to guarantee the work finishes.

See webext-redux's "aliased" action creators:
https://github.com/tshaddix/webext-redux/tree/95ff156b4afe9bfa697e55bfdb32ec116706aba3#4-optional-implement-actions-whose-logic-only-happens-in-the-background-script-we-call-them-aliases