Ensure dead service worker is recoverable
Should the service worker (in Chrome) become responsive for any reason,
like because of the the bug https://crbug.com/1316588, use an [alarm][0]
to periodically restart it. In the catastrophic case where the user
opens the popup, the worker is dead, and the alarm is not going to
restart it soon, give the user a button to reload the entire extension.

[0]: https://developer.chrome.com/docs/extensions/reference/alarms/