Update tab info correctly for `onReplaced` event
The `onReplaced` event is fired when a tab is discarded, which is
something that happens frequently in both Firefox and in Chrome. Update
the tab's ID in the list of locked IDs as well as in the list of tab
times to correctly maintain state after `onReplaced`.

This also stops counting "discarded" tabs as tabs closed tabs, which may
have reduced Tab Wrangler close rate by incorrectly incrementing
denominator of (tabs closed by tab wrangler / all tabs closed).

Closes #462
Closes #468